Remove pro-tip about the CLI (#64)
As I remembered, the reason we recommend Now Desktop purely because of the auto-updating CLI.
If we go with mentioned CLI, they don't comes with that even a installer.
So, I'd like to remove it from here. 

The idea of five-minute guide to let the users to start using Now ASAP.
So, this is not needed.