Update for readability
If we are going to use the Tangle everywhere to refer to an IOTA network, we should be consistent. The last time I asked Lewis, he said that the Tangle is used to refer only to the mainnet network, in which case we shouldn't use the Tangle here, as the Compass is for creating private IOTA networks.

We should also be consistent with how we refer to transactions being sent to through the network. Are they sent through an IOTA network, published to an IOTA network, propagated through an IOTA network. Personally, we should use Plain English and say "send" when a client sends a transaction to a node (active voice). "Send a transaction to an IRI node", we should say propagate when we refer to the act of transactions being sent through each IRI node in a network (passive) "propagate through the network".

If we mention the acronym for the Coodinator (COO), we should use the acronym anywhere else to refer to the Coordinator.

Are we actively not using articles with Compass? Why do we not say "the compass". We say "the Tangle", "the Coodinator", "the IRI". It seems strange and inconsistent.