Update simple-install.md
Made the titles more consistent. All task titles should begin with a gerund if we are to be consistent.

We should avoid using "we" everywhere. We are telling the readers what they should do, so we should address them personally by using "you". I think we can use we in the introductory paragraph, as in "In this guide, we use Ubuntu blah blah". 

To make things a bit clear, I stripped down each step to the bare essentials and added step numbers.

If we need to include some information that is related to a step, we should use a **Note:** paragraph.

I added some links that might be useful for readers.

I was a bit confused by the Merkle tree section, so I separated the content into paragraphs and tried to rephrase some sentences.